api/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
python-jose[cryptography]==3.4.0
httpx==0.28.1
